requirements.txt
notebooklm-py>=0.3.4
notion-client>=2.2.0
python-dotenv>=1.0.0
requests>=2.31.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-telegram-bot>=21.0
